Removed Arithmetic namespace
