+++ katas/content/multi_qubit_measurements/partial_measurements_demo.qs
@@ -1,6 +1,5 @@
 namespace Kata {
     open Microsoft.Quantum.Diagnostics;
-    open Microsoft.Quantum.Arithmetic;
     open Microsoft.Quantum.Convert;
     open Microsoft.Quantum.Math;
