+++ library/std/canon.qs
@@ -515,4 +515,56 @@ namespace Microsoft.Quantum.Canon {
         }
     }
 
+    /// # Summary
+    /// Applies a bitwise-XOR operation between a classical integer and an
+    /// integer represented by a register of qubits.
+    ///
+    /// # Description
+    /// Applies `X` operations to qubits in a little-endian register based on
+    /// 1 bits in an integer.
+    ///
+    /// Let us denote `value` by a and let y be an unsigned integer encoded in `target`,
+    /// then `ApplyXorInPlace` performs an operation given by the following map:
+    /// |y⟩ ↦ |y ⊕ a⟩, where ⊕ is the bitwise exclusive OR operator.
+    operation ApplyXorInPlace(value : Int, target : Qubit[]) : Unit is Adj + Ctl {
+        body (...) {
+            Fact(value >= 0, "`value` must be non-negative.");
+            mutable runningValue = value;
+            for q in target {
+                if (runningValue &&& 1) != 0 {
+                    X(q);
+                }
+                set runningValue >>>= 1;
+            }
+            Fact(runningValue == 0, "value is too large");
+        }
+        adjoint self;
+    }
+
+    /// # Summary
+    /// Applies a bitwise-XOR operation between a classical integer and an
+    /// integer represented by a register of qubits.
+    ///
+    /// # Description
+    /// Applies `X` operations to qubits in a little-endian register based on
+    /// 1 bits in an integer.
+    ///
+    /// Let us denote `value` by a and let y be an unsigned integer encoded in `target`,
+    /// then `ApplyXorInPlace` performs an operation given by the following map:
+    /// |y⟩ ↦ |y ⊕ a⟩, where ⊕ is the bitwise exclusive OR operator.
+    operation ApplyXorInPlaceL(value : BigInt, target : Qubit[]) : Unit is Adj + Ctl {
+        body (...) {
+            Fact(value >= 0L, "`value` must be non-negative.");
+            mutable runningValue = value;
+            for q in target {
+                if (runningValue &&& 1L) != 0L {
+                    X(q);
+                }
+                set runningValue >>>= 1;
+            }
+            Fact(runningValue == 0L, "`value` is too large.");
+        }
+        adjoint self;
+    }
+
 }
