+++ library/std/measurement.qs
@@ -4,6 +4,7 @@
 namespace Microsoft.Quantum.Measurement {
     open Microsoft.Quantum.Core;
     open Microsoft.Quantum.Intrinsic;
+    open Microsoft.Quantum.Diagnostics;
     open QIR.Intrinsic;
 
     /// # Summary
@@ -129,4 +130,35 @@ namespace Microsoft.Quantum.Measurement {
     operation MResetZ (target : Qubit) : Result {
         __quantum__qis__mresetz__body(target)
     }
+
+    /// # Summary
+    /// Measures the content of a quantum register and converts
+    /// it to an integer. The measurement is performed with respect
+    /// to the standard computational basis, i.e., the eigenbasis of `PauliZ`.
+    ///
+    /// # Input
+    /// ## target
+    /// A quantum register in the little-endian encoding.
+    ///
+    /// # Output
+    /// An unsigned integer that contains the measured value of `target`.
+    ///
+    /// # Remarks
+    /// This operation resets its input register to the |00...0> state,
+    /// suitable for releasing back to a target machine.
+    @Config(Full)
+    operation MeasureInteger(target : Qubit[]) : Int {
+        let nBits = Length(target);
+        Fact(nBits < 64, $"`Length(target)` must be less than 64, but was {nBits}.");
+
+        mutable number = 0;
+        for i in 0..nBits-1 {
+            if (MResetZ(target[i]) == One) {
+                set number |||= 1 <<< i;
+            }
+        }
+
+        number
+    }
+
 }
