+++ library/tests/src/resources/compare.qs
@@ -1,6 +1,5 @@
 namespace Test {
     open Microsoft.Quantum.Measurement;
-    open Microsoft.Quantum.Arithmetic;
     open Microsoft.Quantum.Unstable.Arithmetic;
     open Microsoft.Quantum.Convert;
     open Microsoft.Quantum.Diagnostics;
