+++ library/tests/src/resources/inc_by_le.qs
@@ -1,8 +1,8 @@
 namespace Test {
-    open Microsoft.Quantum.Arithmetic;
     open Microsoft.Quantum.Unstable.Arithmetic;
     open Microsoft.Quantum.Convert;
     open Microsoft.Quantum.Diagnostics;
+    open Microsoft.Quantum.Measurement;
 
     internal operation TestIncByLE2(
         name : String,
