+++ library/tests/src/resources/select.qs
@@ -1,5 +1,4 @@
 namespace Test {
-    open Microsoft.Quantum.Arithmetic;
     open Microsoft.Quantum.Arrays;
     open Microsoft.Quantum.Convert;
     open Microsoft.Quantum.Diagnostics;
