+++ samples/algorithms/Shor.qs
@@ -10,7 +10,6 @@ namespace Sample {
     open Microsoft.Quantum.Diagnostics;
     open Microsoft.Quantum.Random;
     open Microsoft.Quantum.Math;
-    open Microsoft.Quantum.Arithmetic;
     open Microsoft.Quantum.Unstable.Arithmetic;
     open Microsoft.Quantum.Arrays;
